Change how popular major versions determined (#26)
See #25

The current methods of determining popularity is to sum, then rank, all polled download rates over the full window where a chart shows versions. This effectively tries to show the versions with the most volume throughout the entire chart.

When we release a new version, we're more interested in what happens in a recent time-period, e.g. the most recent week, as npm is able to most granularly measure.

This biasing doesn't work well for prerelease versions, which may not have downloads for more than a couple of days to use as ranking, but the approach of ranking based on most recent polling will show what we want for longer-lived versions, such as new major releases.

0.68 shows up after this change, though there's still a data spike issue.