Update minimum versions to 0.66ish
This change makes it so that we don't include data from before React Native 0.66 (~1y ago) in the bundle, and do not consider it eligible in ranking. We've continues to see abnormal spikes in the samples directly from NPM, as versions get older, so this should make the good data much more usable. So that the charts are consistent with each other. Projects outside of RN versioning were aligned to their release closest to the time RN/RNW released 0.66.