Add Prettier command
Prettier was added, but configured halfway.

This PR adds `yarn prettier` and `yarn prettier:fix` commands, and ensures auto generated files will not be touched by Prettier.

It also formats all files using Prettier so that the command can actually be used to validate if everything is... Pretty.

I skipped formatting of files which were heavily modified in #34 to avoid conflicts.