Remove __dirname altogether
Since scripts are to be executed using Yarn-manages scripts anyway, __dirname workaround is not really relevant, because wherever we would actually execute the scripts from, "root" is always going to be where we expect it to be. So, we don't need paths to be relative to the scripts - we can make them relative to the root instead.