Trim trailing `-1`s when generating assets (#70)
* Trim trailing -1s when generating assets

In case versionCounts are stored as an array (something that landed as an optimization in #40), we then read the array by skipping all array items with `-1`s. While this is already more optimal than an object in many cases, I've noticed that we could just trim trailing chains of `-1`s in this array, not only reducing the amount of data, but also making it faster to parse.

generated_assets directory size
before: 3 840 513 B
after: 3 707 218 B (-3.5%)

* Even simpler while() loop